add confirm to delete

+++ notes.php
@@ -82,7 +82,7 @@ if ($_POST) {
                     <td><?php echo $data[1] ?></td>
                     <td><?php echo $data[3] ?></td>
                     <td class="text-center"><a href="edit-note.php?id=<?php echo $data[0]?>"><i class="far fa-edit"></i></a></td>
-                    <td class="text-center"><a href="delete.php?id=<?php echo $data[0]?>&which=note"><i class="far fa-trash-alt"></i></a></td>
+                    <td class="text-center"><a href="delete.php?id=<?php echo $data[0]?>&which=note" onclick="return confirm('Do you want to delete this note? Y/N')"><i class="far fa-trash-alt"></i></a></td>
                   </tr>              
                 <?php } ?>          
               </tbody>
